+++ packages/match-list/demo/session.js
@@ -1,0 +1,14 @@
+module.exports = [
+  {
+    id: '1',
+    value: [1, 4, undefined, undefined]
+  },
+  {
+    id: '2',
+    value: [1, 2, 3, 4]
+  },
+  {
+    id: '3',
+    value: [1, 2, 3, 4]
+  }
+];