+++ packages/math-inline/src/index.js
@@ -25,6 +25,7 @@ export default class Match extends HTMLElement {
   sessionChanged(s) {
     this._session.answers = s.answers;
     this._session.response = s.response;
+    this._session.completeAnswer = s.completeAnswer;
     log('session: ', this._session);
   }