+++ packages/placement-ordering/src/main.jsx
@@ -1,12 +1,16 @@
 import PlacementOrdering from './placement-ordering';
 import React from 'react';
 import { withDragContext } from '@pie-lib/drag';
+
 import PropTypes from 'prop-types';
 
 class Main extends React.Component {
   static propTypes = {
     model: PropTypes.object.isRequired,
-    session: PropTypes.object.isRequired,
+    session: PropTypes.oneOfType([
+      PropTypes.array.isRequired,
+      PropTypes.object.isRequired
+    ]),
     onSessionChange: PropTypes.func.isRequired
   };
 
@@ -21,5 +25,5 @@ class Main extends React.Component {
     );
   }
 }
-
-export default withDragContext(Main);
+const out = withDragContext(Main);
+export default out;