+++ packages/multiple-choice/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.14](https://github.com/pie-framework/pie-elements/compare/@pie-ui/multiple-choice@4.7.13...@pie-ui/multiple-choice@4.7.14) (2019-04-25)
+
+**Note:** Version bump only for package @pie-ui/multiple-choice
+
+
+
+
+
+## [4.7.13](https://github.com/pie-framework/pie-elements/compare/@pie-ui/multiple-choice@4.7.12...@pie-ui/multiple-choice@4.7.13) (2019-04-25)
+
+**Note:** Version bump only for package @pie-ui/multiple-choice
+
+
+
+
+
+## [4.7.12](https://github.com/pie-framework/pie-elements/compare/@pie-ui/multiple-choice@4.7.11...@pie-ui/multiple-choice@4.7.12) (2019-04-18)
+
+**Note:** Version bump only for package @pie-ui/multiple-choice
+
+
+
+
+
+## [4.7.11](https://github.com/pie-framework/pie-elements/compare/@pie-ui/multiple-choice@4.7.10...@pie-ui/multiple-choice@4.7.11) (2019-04-18)
+
+
+### Bug Fixes
+
+* Updated @pie-lib/config-ui and @pie-lib/render-ui versions. ([1092ed3](https://github.com/pie-framework/pie-elements/commit/1092ed3))
+
+
+
+
+
+## [4.7.10](https://github.com/pie-framework/pie-elements/compare/@pie-ui/multiple-choice@4.7.9...@pie-ui/multiple-choice@4.7.10) (2019-04-15)
+
+**Note:** Version bump only for package @pie-ui/multiple-choice
+
+
+
+
+
+## [4.7.9](https://github.com/pie-framework/pie-elements/compare/@pie-ui/multiple-choice@4.7.8...@pie-ui/multiple-choice@4.7.9) (2019-04-10)
+
+**Note:** Version bump only for package @pie-ui/multiple-choice
+
+
+
+
+
+## [4.7.8](https://github.com/pie-framework/pie-elements/compare/@pie-ui/multiple-choice@4.7.7...@pie-ui/multiple-choice@4.7.8) (2019-04-09)
+
+**Note:** Version bump only for package @pie-ui/multiple-choice
+
+
+
+
+
 ## [4.7.7](https://github.com/pie-framework/pie-elements/compare/@pie-ui/multiple-choice@4.7.6...@pie-ui/multiple-choice@4.7.7) (2019-04-04)
 
 **Note:** Version bump only for package @pie-ui/multiple-choice